Graph Solvers

Supported Solvers

NavAbility provides both parametric and nonparametric (i.e. non-Gaussian / multi-modal) factor graph solving solutions. Please reach out via email <info@navability.io> or Slack if you would like to learn more about these features.

Supported Probabily Models

Gaussian Case

Parametric solving supports measurement models using the probability beliefs:

Normal

MvNormal

And also currently in beta:

Max-Mixture of Normal or MvNormal

Non-Gaussian Case

Non parametric solving can support a much wider range of measurement probability models. Currently the SDK supports

Mixture

Uniform

Rayleigh

as well as multihypothesis features per factor – i.e. Categorical.

Many more probability types are natively supported by the solver and yet directly exposed through the SDK, including

Manifold Kernel Densities,

Heatmaps or Intensity maps,

Levelsets,

Most common parametric probability models.

Reach out to NavAbility if you seek earlier SDK support for these or other more advanced probability features.

Probability Model Index

NavAbilitySDK.Normal — Type

mutable struct Normal <: Distribution

One dimensional normal distribution.

NavAbilitySDK.FullNormal — Type

mutable struct FullNormal <: Distribution

Multidimensional normal distribution specified by means and a covariance matrix.

NavAbilitySDK.Uniform — Type

mutable struct Uniform <: Distribution

One dimensional uniform distribution.

NavAbilitySDK.Rayleigh — Type

mutable struct Rayleigh <: Distribution

One dimensional Rayleigh distribution.

NavAbilitySDK.MixtureData — Function

MixtureData(mechanics, components, probabilities, dims)

Create a Mixture factor type with an underlying factor type, a named set of distributions that should be mixed, the probabilities of each distribution (the mix), and the dimensions of the underlying factor (e.g. ContinuousScalar=1, Pose2Pose2=3, etc.).

Args: mechanics (Type{FactorData}): The underlying factor data type, e.g. Pose2Pose2Data. NOTE: This will change in later versions but for now it can be any of the FactorData classes (e,g, LinearRelative, not the object LinearRelative()). components (NamedTuple): The named tuple set of distributions that should be mixed, e.g. NamedTuple(hypo1=Normal(0, 2)), hypo2=Uniform(30, 55)). probabilities (List[float]): The probabilities of each distribution (the mix), e.g. [0.4, 0.6]. dims (int): The dimensions of the underlying factor, e.g. for Pose2Pose2 it's 3.