SDK function Index

Documentation for NavAbilitySDK.

NavAbilitySDK.Categorical

NavAbilitySDK.ContinuousScalar

NavAbilitySDK.Distribution

NavAbilitySDK.FullNormal

NavAbilitySDK.MixtureInferenceType

NavAbilitySDK.Normal

NavAbilitySDK.Point2

NavAbilitySDK.Point3

NavAbilitySDK.Pose2

NavAbilitySDK.Pose2AprilTag4Corners

NavAbilitySDK.Pose2Point2BearingRange

NavAbilitySDK.Pose3

NavAbilitySDK.Rayleigh

NavAbilitySDK.ScatterAlignPose2

NavAbilitySDK.Uniform

DistributedFactorGraphs.addVariable!

DistributedFactorGraphs.getId

NavAbilitySDK.createConnect

NavAbilitySDK.createUpload

NavAbilitySDK.@nvaZInferenceType

NavAbilitySDK.Categorical — Type

mutable struct Categorical <: NavAbilitySDK.Distribution

Categorical distribution specified by a set of probabilities summing up to 1.

NavAbilitySDK.ContinuousScalar — Type

struct ContinuousScalar <: NavAbilitySDK.VariableType

General 1 dimensional variable.

NavAbilitySDK.Distribution — Type

abstract type Distribution

Abstract parent type for all distributions.

NavAbilitySDK.FullNormal — Type

mutable struct FullNormal <: NavAbilitySDK.Distribution

Multidimensional normal distribution specified by means and a covariance matrix.

NavAbilitySDK.MixtureInferenceType — Type

struct MixtureInferenceType <: DistributedFactorGraphs.InferenceType

InferenceType for MixtureData.

NavAbilitySDK.Normal — Type

mutable struct Normal <: NavAbilitySDK.Distribution

One dimensional normal distribution.

NavAbilitySDK.Point2 — Type

struct Point2 <: NavAbilitySDK.VariableType

XY Euclidean manifold variable.

NavAbilitySDK.Point3 — Type

struct Point3 <: NavAbilitySDK.VariableType

XYZ Euclidean manifold variable..

NavAbilitySDK.Pose2 — Type

struct Pose2 <: NavAbilitySDK.VariableType

SpecialEuclidean(2) manifold, two Euclidean translations and one Circular rotation, used for general 2D SLAM.

NavAbilitySDK.Pose2AprilTag4Corners — Type

struct Pose2AprilTag4Corners <: DistributedFactorGraphs.InferenceType

InferenceType for Pose2AprilTag4CornersData.

NavAbilitySDK.Pose2Point2BearingRange — Type

struct Pose2Point2BearingRange <: DistributedFactorGraphs.InferenceType

Pose2Point2BearingRangeInferenceType is used to represent a bearing

range measurement.

NavAbilitySDK.Pose3 — Type

struct Pose3 <: NavAbilitySDK.VariableType

SpecialEuclidean(3) manifold used for 3D SLAM.

NavAbilitySDK.Rayleigh — Type

mutable struct Rayleigh <: NavAbilitySDK.Distribution

One dimensional Rayleigh distribution.

NavAbilitySDK.ScatterAlignPose2 — Type

Alignment factor between point cloud populations, using either

a continuous density function cost: ApproxManifoldProducts.mmd, or

a conventional iterative closest point (ICP) algorithm (when .sample_count < 0).

This factor can support very large density clouds, with sample_count subsampling for individual alignments.

NavAbilitySDK.Uniform — Type

mutable struct Uniform <: NavAbilitySDK.Distribution

One dimensional uniform distribution.

DistributedFactorGraphs.addVariable! — Method

addVariable! Add a variable to the NavAbility Platform service Example

addVariable!(fgclient, "x0", NvaSDK.Pose2)

DistributedFactorGraphs.getId — Method

getId

Get the deterministic identifier (uuid v5) for a node.

NavAbilitySDK.createConnect — Method

createConnect

Create a connection gql query to a node.

NavAbilitySDK.createUpload — Function

createUpload(nvastore, blobId)
createUpload(nvastore, blobId, parts)

Request URLs for data blob upload.

Args: navAbilityClient (NavAbilityClient): The NavAbility client. blobId: The unique file identifier of the data blob. parts (Int): Split upload into multiple blob parts, FIXME currently only supports parts=1.

NavAbilitySDK.@nvaZInferenceType — Macro

Macro to autogenerate factor type definitions. Used for common types such as Prior, PriorPose2, Pose3Pose3, etc.